+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## master (unreleased)
 
+* [#333](https://github.com/clojure-emacs/orchard/pull/333): Add `orchard.profile`.
+
 ## 0.32.1 (2025-04-05)
 
 * [#328](https://github.com/clojure-emacs/orchard/pull/328): Inspector: display identity hashcode for Java objects.